moved S/MIME initialization into msg and the pointer to the data structure, also adjusted tests
